Fix erroneous check when saving EXPRESSION_DATA_FLOW_INFO
Before saving a data flow info, we checked if it's the same as that of context,
which was passed to us. This proves wrong when eventually somebody asks for a
data flow info of the same expression, but with a different context (i.e. with
a different data flow info in it) and we return this (different) context's
info, whereas we should have returned that of the original context.
This could've happened in CandidateResolver, for example.

Save data flow info after every expression now, only if it's not empty.

No new tests added, see next commit.