Skip Plugin Exceution for POM Projects
It's quite common to have a multi module project with a parent project
that has POM packaging. It would be quite convenient if the
kotlin-maven-plugin could be configured there instead of in each module.
Currently this is not possible because the plugin tries to execute on
POM projects with fails.

 - skip execution on POM projects