Prevent annoying tooltip in Kotlin project
Module 'builtins' has no dependency on kotlin-runtime.jar, yet it has a lot of
Kotlin files. Add jet.Unit to the list of marker classes as an example of the
class not from the stdlib, but from built-ins. Since 'builtins' has jet.Unit in
its classpath, no tooltip is displayed